[SCRUM-95] refactor: 서비스 계층 인터페이스 제거
